+++ Changes.md
@@ -28,11 +28,13 @@ Improvements
 ------------
 
 - RenderMan : Stylized Looks no longer require manual AOV setup. The relevant AOVs are added automatically whenever a stylized display filter is present.
+- Arnold : Added translation of UsdLux IES parameters to Arnold.
 
 Fixes
 -----
 
 - RenderManShader : Fixed handling of minimum and maximum values for `color`, `vector`, `normal` and `point` parameters.
+- FreezeTransform : Fixed double transforming when the input primitive contains multiple primitive variables sharing the same data.
 
 1.5.10.1 (relative to 1.5.10.0)
 ========